删除分类

+++ server/api/admin/sort/deleteSort.js
@@ -1,5 +1,6 @@
 const chalk = require('chalk')
 const sortUtils = require('../../../mongodb/utils/sorts')
+const postUtils = require('../../../mongodb/utils/posts')
 const utils = require('../../../utils/utils')
 const log4js = require('log4js')
 const adminApiLog = log4js.getLogger('adminApi')
@@ -14,13 +15,16 @@ module.exports = async function (req, res, next) {
     })
     return
   }
-  // TODO:分类下有文章时，不允许删除
-  res.status(400).json({
-    errors: [{
-      message: 'TODO:分类下有文章时，不允许删除'
-    }]
-  })
-  return
+  // 分类下有文章时，不允许删除
+  const post = await postUtils.findOne({ sort: id })
+  if (post) {
+    res.status(400).json({
+      errors: [{
+        message: '该分类下有文章，不允许删除'
+      }]
+    })
+    return
+  }
   //  删除分类
   sortUtils.deleteOne({ _id: id }).then((data) => {
     if (data.deletedCount === 0) {
